authorization_service/requirements.txt
werkzeug
tmdbsimple
celery
redis
postgres
psycopg2-binary
Flask-SQLAlchemy
requests
sendgrid
flask
SQLAlchemy
pyjwt
